// apps/chats-service/src/generated.grpc/chats.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.7
//   protoc               v3.19.1
// source: chats.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { wrappers } from "protobufjs";
import { Observable } from "rxjs";
import { Struct } from "./google/protobuf/struct";
import { Timestamp } from "./google/protobuf/timestamp";

export const protobufPackage = "poslah.chats";

/** The request message for the GetLastMessageEvents method. */
export interface GetLastMessageEventsRequest {
  /** The UUID of the chat to fetch events for. */
  chatId: string;
  /** The `nn` of the message for which we are getting events. */
  messageNn: bigint;
  /** The event `nn` after which to start fetching. */
  afterNn: bigint;
}

/** A single event, containing its number and payload. */
export interface LastMessageEvent {
  /** The event's unique sequence number within the chat. */
  nn: bigint;
  /** The payload of the event (with message partial snapshot). */
  payload: { [key: string]: any } | undefined;
}

/** The response message, which contains a list of the found events. */
export interface GetLastMessageEventsResponse {
  /** createdAt from oldest of returning events (if any returned) */
  oldestCreatedAt?:
    | Timestamp
    | undefined;
  /** An array of LastMessageEvent objects. */
  events: LastMessageEvent[];
}

export const POSLAH_CHATS_PACKAGE_NAME = "poslah.chats";

wrappers[".google.protobuf.Struct"] = { fromObject: Struct.wrap, toObject: Struct.unwrap } as any;

export interface ChatsServiceClient {
  /** A method to get all message-related events for a specific chat after a certain point. */

  getLastMessageEvents(request: GetLastMessageEventsRequest): Observable<GetLastMessageEventsResponse>;
}

export interface ChatsServiceController {
  /** A method to get all message-related events for a specific chat after a certain point. */

  getLastMessageEvents(
    request: GetLastMessageEventsRequest,
  ): Promise<GetLastMessageEventsResponse> | Observable<GetLastMessageEventsResponse> | GetLastMessageEventsResponse;
}

export function ChatsServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["getLastMessageEvents"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("ChatsService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("ChatsService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const CHATS_SERVICE_NAME = "ChatsService";